refactor: some code consistency changes
* changes to led api to be consistent with other functional apis
* change console log to be more informative in tests
* let typescript assume a functions return type instead of setting it explicitly
