+++ src/led/ffi-lib.ts
