+++ src/tests/led/run-tests.ts
@@ -5,7 +5,7 @@ import * as functionalTests from './functional-api';
 
 export async function runLedTests()
 {
-	console.log('Old API');
+	console.log('Functional API');
 	console.log('---');
 
 	await executeTest('getting the sdk version', functionalTests.testColor);
